+++ src/main/java/dev/wroud/mc/worlds/mixin/TheEndGatewayBlockEntityAccessor.java
@@ -3,7 +3,7 @@ package dev.wroud.mc.worlds.mixin;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.block.entity.TheEndGatewayBlockEntity;
-import net.minecraft.world.level.levelgen.feature.configurations.EndGatewayConfiguration;
+import net.minecraft.world.level.levelgen.feature.EndGatewayFeature;
 import org.spongepowered.asm.mixin.Mixin;
 import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.gen.Invoker;
@@ -17,7 +17,7 @@ public interface TheEndGatewayBlockEntityAccessor {
     }
     
     @Invoker("spawnGatewayPortal")
-    static void invokeSpawnGatewayPortal(ServerLevel serverLevel, BlockPos blockPos, EndGatewayConfiguration endGatewayConfiguration) {
+    static void invokeSpawnGatewayPortal(ServerLevel serverLevel, BlockPos blockPos, EndGatewayFeature feature) {
         throw new AssertionError("Mixin invoker should not be called directly");
     }